--- FIlereading.py
@@ -1,3 +1,0 @@
-import numpy as np
-arr = np.load("C:\\Users\\hayde\\OneDrive\\Desktop\\images\\HeNe_darks_1200ms_2025-01-24_16-01-20.npz")
-print(arr['array'])
